feat(types): add optional HMPB `layout` property
Allows interpreters to describe the layout of the ballot image. Moves some types & schemas around to prevent dependency cycles.
